added new util to compute section entropie; minor refactors
