+++ obfuscation_detection/heuristics.py
@@ -1,6 +1,6 @@
 from binaryninja import highlight
 
-from .utils import *
+from .helpers import *
 
 
 def find_flattened_functions(bv):
@@ -124,26 +124,3 @@ def find_complex_arithmetic_expressions(bv):
         if score != 0:
             print(
                 f"Function {hex(f.start)} ({(f.name)}) has {score} instructions that use complex arithmetic expressions.")
-
-
-def find_entry_functions(bv):
-    print("=" * 80)
-    print("Functions without callers:")
-
-    for f in bv.functions:
-        if len(f.callers) != 0:
-            continue
-
-        print(
-            f"Function {hex(f.start)} ({(f.name)}) has no known callers.")
-
-
-def find_leaf_functions(bv):
-    print("=" * 80)
-    print("Functions without callees:")
-
-    for f in bv.functions:
-        # no callees and at least two instructions
-        if len(f.callees) == 0 and sum(1 for _ in f.instructions) > 1:
-            print(
-                f"Function {hex(f.start)} ({(f.name)}) has no known callees.")
